server/requirements.txt
fastapi
uvicorn
langchain-core
langchain_openai
python-dotenv
pydantic
moviepy
azure-storage-blob
urllib3